web_app_with_drf_test/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==5.0.1
django-cacheops==7.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-oauth-toolkit==2.3.0
django-redis==5.4.0
django-rest-framework-social-oauth2==1.2.0
django-social-share==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
funcy==2.0
idna==3.6
jwcrypto==1.5.1
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wrapt==1.16.0
